Refactored Dialog to Fragment behavior and added new screens.
