--- togetherbutton/src/main/java/com/github/togetherproject/button/helpServices/ServicesAdapter.kt
@@ -1,4 +1,0 @@
-package com.github.togetherproject.button.helpServices
-
-class ServicesAdapter {
-}
