--- togetherbutton/src/main/java/com/github/togetherproject/button/helpServices/ServicesDAO.kt
@@ -1,4 +1,0 @@
-package com.github.togetherproject.button.dao
-
-class ServicesDAO: IServicesDAO {
-}
