+++ togetherbutton/src/main/res/layout/fragment_together_bottom_sheet.xml
@@ -1,51 +1,25 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/bottomSheetConstraint"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
+    android:background="@drawable/background_bottom_sheet"
+    android:orientation="vertical"
     android:paddingStart="8dp"
     android:paddingTop="8dp"
     android:paddingEnd="8dp"
     android:paddingBottom="24dp"
-    android:background="@drawable/background_bottom_sheet"
     app:layout_behavior="@string/bottom_sheet_behavior">
 
-    <ImageView
-        android:id="@+id/imgViewTopLogo"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="12dp"
-        android:contentDescription="@string/content_together_top_logo"
-        android:src="@drawable/ic_logo_together"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageButton
-        android:id="@+id/imgBtnClose"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="18dp"
-        android:background="@drawable/ic_circular_close"
-        android:contentDescription="@string/content_close_button"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="@id/imgViewTopLogo" />
+    <include
+        android:id="@+id/headerLayout"
+        layout="@layout/together_header" />
 
     <FrameLayout
-        android:id="@+id/screens"
-        android:layout_width="0dp"
+        android:id="@+id/fragmentContainer"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_marginTop="18dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imgViewTopLogo">
+        android:layout_marginTop="18dp" />
 
-        <include
-            android:id="@+id/layout_box"
-            layout="@layout/screen_first" />
-    </FrameLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
+</LinearLayout>
