--- togetherbutton/src/main/res/layout/screen_first.xml
@@ -1,72 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content">
-
-    <TextView
-        android:id="@+id/txtViewMainText3"
-        style="@style/headerText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        android:text="@string/home_we_need_to_talk"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <ImageView
-        android:id="@+id/imgViewCenter2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:contentDescription="@string/content_main_center_img"
-        android:scaleType="fitCenter"
-        android:src="@drawable/ic_man_woman_phone"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtViewMainText3" />
-
-    <TextView
-        android:id="@+id/txtViewHumanHint2"
-        style="@style/secondaryText"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="18dp"
-        android:gravity="center"
-        android:text="@string/find_help"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/imgViewCenter2" />
-
-    <LinearLayout
-        android:id="@+id/llActionButtons"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:gravity="center"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtViewHumanHint2">
-
-        <Button
-            android:id="@+id/btnCall"
-            style="@style/primaryButtonText"
-            android:layout_width="145dp"
-            android:layout_height="wrap_content"
-            android:background="@drawable/background_main_button"
-            android:foreground="?android:selectableItemBackground"
-            android:text="@string/ask_for_help"
-            android:textAllCaps="false"
-            android:textColor="@color/colorWhite" />
-
-        <Button
-            android:id="@+id/btnContacts"
-            style="@style/primaryButtonText"
-            android:layout_width="145dp"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:background="@drawable/background_main_button"
-            android:text="@string/safe_contacts"
-            android:textAllCaps="false"
-            android:textColor="@color/colorWhite" />
-    </LinearLayout>
-</androidx.constraintlayout.widget.ConstraintLayout>
